pyproject.toml
[project]
name = "headless-factory"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.10"
dependencies = [
    "google-genai>=1.68.0",
    "google-generativeai>=0.8.6",
    "openai>=2.29.0",
    "python-dotenv>=1.2.2",
]
